pyproject.toml
[project]
name = "web3-mcp"
dynamic = ["version"]
description = "FastMCP server wrapping Ankr Advanced API"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ankr-sdk>=1.0.2",
    "fastmcp>=2.2.0",
]

[project.scripts]
web3-mcp = "web3_mcp.__main__:main"

[project.optional-dependencies]
dev = [
    "black>=25.1.0",
    "isort>=6.0.1",
    "mcp>=1.6.0",
    "mypy>=1.15.0",
    "pre-commit>=4.2.0",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pytest-cov>=5.0.0",
    "ruff>=0.11.5",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 100

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.pytest.ini_options]
asyncio_mode = "strict"
asyncio_default_fixture_loop_scope = "session"

[tool.hatch.version]
path = "src/web3_mcp/__version__.py"
pattern = '__version__ = "(?P<version>.+)"'

[dependency-groups]
dev = [
    "pytest-cov>=7.0.0",
]
